Move `Shape` to dedicated module
